requirements.txt
asgiref==3.3.4
attrs==21.2.0
autopep8==1.5.7
beautifulsoup4==4.9.3
coverage==5.5
Django==3.2.2
django-environ==0.4.5
djangorestframework==3.12.4
gunicorn==20.1.0
iniconfig==1.1.1
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
pytest-cov==2.11.1
pytest-django==4.2.0
pytz==2021.1
selenium==3.141.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.4
